Fix Clang ambiguous setChild overload resolution
StringIndex::PrefixTreeNode has two setChild overloads:
  setChild(PrefixTreeNode*, size_t idx)
  setChild(PrefixTreeNode*, char c)

Passing a uint64_t childSlot is ambiguous under Clang (both implicit
conversions are equally valid) while GCC quietly picked the size_t
overload. Cast the index to size_t to disambiguate.

+++ storage/dump/GraphLoader.cpp
@@ -644,7 +644,7 @@ void GraphLoader::setStringPropIndexChild(EntityKind kind,
     auto& idx = currentStringIndex(kind, propTypeID);
     auto* parent = idx.getNode(parentLocalID);
     auto* child = idx.getNode(childLocalID);
-    parent->setChild(child, childSlot);
+    parent->setChild(child, static_cast<size_t>(childSlot));
 }
 
 void GraphLoader::appendStringPropIndexOwners(EntityKind kind,
